feat: adds user query and resolver
