+++ server/resolvers/mutations.js
@@ -11,7 +11,7 @@ module.exports = {
     // Successful if no errors
     return { success };
   },
-  login: (_, args, { user }) => {
-    // add logic for loging a user in here
-  }
+  // login: (_, args, { user }) => {
+  //   // add logic for loging a user in here
+  // }
 };
