+++ server/resolvers/queries.js
@@ -1,19 +1,9 @@
-const books = [ // TODO: delete me plz
-  {
-    title: "The Awakening",
-    author: "Kate Chopin",
-  },
-  {
-    title: "City of Glass",
-    author: "Paul Auster",
-  },
-  {
-    title: "Harry Potter",
-    author: "J.K. Rowling",
-  },
-];
+const { books, mockUsers } = require('../data/userData')
 
 module.exports = {
   // TODO: delete me as soon as we have a real query
   books: () => books,
+  user: (_, { id }) => {
+    return mockUsers.find(user => user.id === id);
+  }
 };
